Add files via upload
 The global population of electric vehicles (EVs) was growing steadily but still represented a relatively small percentage of the total number of vehicles on the road. Here are some key data points:

According to the International Energy Agency (IEA), the global stock of electric passenger cars surpassed 10 million in 2020, an increase of 41% compared to the previous year. This represents about 1.5% of the total number of passenger cars worldwide.

China was the largest market for electric cars, accounting for around 44% of global electric car sales in 2020. Europe was the second-largest market, with around 28% of sales, followed by North America at 26%.

In terms of individual countries, Norway had the highest EV adoption rate, with electric cars accounting for more than 75% of new car sales in 2020. Other countries with high adoption rates included Iceland, Sweden, and the Netherlands.

The most popular electric car models in 2020 were the Tesla Model 3, the Renault Zoe, and the Volkswagen ID.3, according to data from EV Sales Blog.

Electric buses and two-wheelers (such as e-bikes and electric scooters) were also gaining popularity in many parts of the world. China was the largest market for electric buses, accounting for around 98% of global sales in 2020, while Europe led in e-bike sales.

It's worth noting that the population of electric vehicles is likely to have continued growing since my knowledge cutoff date, as governments and automakers around the world have set ambitious targets for transitioning to electric mobility.